Changed structure, updated test data and added description
