+++ scout/ext/backend/utils/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from __future__ import print_function, absolute_import, unicode_literals
+from .get_case import get_institute, get_case
+from .get_genotype import get_genotype
+from .get_transcript import get_transcript
+from .generate_md5_key import generate_md5_key
+from .get_genes import get_genes
+from .get_compounds import get_compounds
+from .get_mongo_variant import get_mongo_variant
