+++ tests/for_clinical_db/1/scout_config_clinical.ini
@@ -6,6 +6,7 @@ human_genome_build = GRCh
 human_genome_version = 37
 institutes = CMMS,
 madeline = /mnt/hds/proj/cust003/develop/analysis/exomes/1/1ped.xml
+coverage_report = /mnt/hds/proj/cust003/develop/analysis/exomes/1/1.coverage.pdf
 [gene_lists]
   [[MIT]]
     name = MIT
