+++ tests/for_clinical_db/P575-coriell/scout_config_research.ini
@@ -6,6 +6,7 @@ human_genome_build = GRCh
 human_genome_version = 37
 institutes = CMMS,
 madeline = /mnt/hds/proj/cust003/develop/analysis/exomes/P575-coriell/P575-coriellped.xml
+coverage_report = /mnt/hds/proj/cust003/develop/analysis/exomes/P575-coriell/P575-coriell.coverage.pdf
 [gene_lists]
   [[Res]]
     name = Res
