+++ tests/vcf_examples/1/scout_config_test_clinical.ini
@@ -6,6 +6,7 @@ human_genome_build = GRCh
 human_genome_version = 37
 institutes = CMMS,
 madeline = tests/vcf_examples/1/1ped.xml
+coverage_report = tests/vcf_examples/1/1.coverage.pdf
 [gene_lists]
   [[MIT]]
     name = MIT
